// github/types/paths/repos/{owner}/{repo}/milestones/{milestone_number}.types.ts
export type issuesGetMilestone = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesGetMilestone_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: milestone;
          };
        };
        examples: {
          default: unknown;
        };
      };
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type issuesUpdateMilestone = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesUpdateMilestone_Path;
    headers: never;
    cookie: never;
    body: {
      /**
       * The title of the milestone.
       */
      title?: string;
      /**
       * The state of the milestone. Either `open` or `closed`.
       * @default "open"
       */
      state?: "open" | "closed";
      /**
       * A description of the milestone.
       */
      description?: string;
      /**
       * The milestone due date. This is a timestamp in [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format: `YYYY-MM-DDTHH:MM:SSZ`.
       * @format date-time
       */
      due_on?: string;
    };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: milestone;
          };
        };
        examples: {
          default: unknown;
        };
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Deletes a milestone using the given milestone number.
 */
export type issuesDeleteMilestone = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: issuesDeleteMilestone_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type issuesGetMilestone_Path = {
  owner: string;
  repo: string;
  milestone_number: number;
};

export type issuesUpdateMilestone_Path = {
  owner: string;
  repo: string;
  milestone_number: number;
};

export type issuesDeleteMilestone_Path = {
  owner: string;
  repo: string;
  milestone_number: number;
};